1.3: Classes Create the following class and methods:
1. A class called FuelPump with at least these attributes (with respective
setters/getters):
a. fuelPrice → Price per liter
b. fuelQuantity → Remaining amount of fuel, in liters
2. It must have the following functions:
a. fillWithPrice( price )
i. Receives: Paid price
ii. Returns: How many liters were pumped, or zero if there is either not enough
fuel or the price was not set.
4
b. fillWithLiters( liters )
i. Receives: Amount of liters to be filled
ii. Returns: Final paid price, or zero if there is either not enough fuel or the price
was not set.
c. setPrice( price ) – changes the specific liter price. Receives the new price.
d. setQuantity( quantity ) – changes the available fuel quantity in the pump.
Receives the new quantity.
3. Behaviour:
a. The fuelQuantity has to be updated after calling the fillWith functions.
b. fuelQuantity and fuelPrice can never be negative
c. fuelQuantity and fuelPrice must be set on the class constructor